🥅 handle cases with no data
