+++ client-portals/pages/admin/dashboard/[slug].tsx
@@ -62,9 +62,16 @@ const MaintainerPage = () => {
   }, [router.query, accepted]);
 
   return loading2 ? (
-    <div className="flex flex-col items-center justify-center w-screen min-h-screen bg-base-blue">
-      <CSSLoader />
-    </div>
+    <>
+      <Layout type="admin">
+        <div className="flex flex-col items-center justify-center">
+          <CSSLoader />
+        </div>
+      </Layout>
+      <div className="fixed bottom-0 w-full">
+        <Footer />
+      </div>
+    </>
   ) : (
     <>
       <Layout type="admin">
